requirements.txt
# Core
pytest==8.4.1
pytest-bdd==8.1.0
pytest-xdist==3.8.0
selenium==4.35.0
webdriver-manager==4.0.2
PyYAML==6.0.2
# Reporting
allure-pytest-bdd==2.15.0
# API Requests
requests==2.32.5